CTexture implementations (data loading not implemented in D3D9 and not tested in D3D11) CEGUI renderer base

+++ branches/Dev/DEM/Src/L1/Render/RenderTarget.h
@@ -19,8 +19,8 @@ typedef Ptr<class CTexture> PTexture;
 
 struct CRenderTargetDesc
 {
-	ushort			Width;
-	ushort			Height;
+	DWORD			Width;
+	DWORD			Height;
 	EPixelFormat	Format;
 	EMSAAQuality	MSAAQuality;
 	bool			UseAsShaderInput;
